refactor: reuse provider auth and diagnostic helpers
