+++ Sources/CodexBarCore/Providers/VertexAI/VertexAIOAuth/VertexAIIDToken.swift
@@ -1,0 +1,28 @@
+import Foundation
+
+/// Reads display identity only; token verification belongs to the OAuth provider.
+enum VertexAIIDToken {
+    static func email(from token: String?) -> String? {
+        guard let token, !token.isEmpty else { return nil }
+
+        let parts = token.components(separatedBy: ".")
+        guard parts.count >= 2 else { return nil }
+
+        var payload = parts[1]
+            .replacingOccurrences(of: "-", with: "+")
+            .replacingOccurrences(of: "_", with: "/")
+
+        let remainder = payload.count % 4
+        if remainder > 0 {
+            payload += String(repeating: "=", count: 4 - remainder)
+        }
+
+        guard let data = Data(base64Encoded: payload, options: .ignoreUnknownCharacters),
+              let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any]
+        else {
+            return nil
+        }
+
+        return json["email"] as? String
+    }
+}
